+++ tests/test_templatetags.py
@@ -11,7 +11,6 @@ def test_tailwind_css_in_production(settings):
     ).render(Context({}))
 
     assert '<link rel="stylesheet" href="/static/css/dist/styles.css">' in output
-    assert "browser-sync/browser-sync-client.js" not in output
 
 
 def test_tailwind_css_in_production_with_version(settings):
@@ -24,7 +23,6 @@ def test_tailwind_css_in_production_with_version(settings):
     ).render(Context({}))
 
     assert '<link rel="stylesheet" href="/static/css/dist/styles.css?v=123">' in output
-    assert "browser-sync/browser-sync-client.js" not in output
 
 
 def test_tailwind_css_in_debug(settings):
@@ -38,7 +36,6 @@ def test_tailwind_css_in_debug(settings):
     ).render(Context({}))
 
     assert '<link rel="stylesheet" href="/static/css/dist/styles.css">' in output
-    assert "//HOST:8383/browser-sync/browser-sync-client.js" in output
 
 
 def test_tailwind_css_in_debug_with_version(settings):
@@ -52,7 +49,6 @@ def test_tailwind_css_in_debug_with_version(settings):
     ).render(Context({}))
 
     assert '<link rel="stylesheet" href="/static/css/dist/styles.css?v=123">' in output
-    assert "//HOST:8383/browser-sync/browser-sync-client.js" in output
 
 
 def test_tailwind_preload_css(settings):